support deploying acr/dockerhub/otherpublic image to container app.
