+++ azure-toolkit-libs/azure-toolkit-containerapps-lib/src/main/java/com/microsoft/azure/toolkit/lib/containerapps/containerapp/ContainerAppModule.java
@@ -54,7 +54,7 @@ public class ContainerAppModule extends AbstractAzResourceModule<ContainerApp, A
     @Nonnull
     @Override
     protected ContainerAppDraft newDraftForUpdate(@Nonnull ContainerApp containerApp) {
-        throw new UnsupportedOperationException("not support");
+        return new ContainerAppDraft(containerApp);
     }
 
     @Nonnull
